Add clang and llvm as submodules, together with a new Makefile
This adds `rsync` as a build dependency, and requires git clone
to be used with the `--recursive` flag. In a preixisting repo,
run `git submodule update --init` in order to clone the
submodules.

The build process is reduced to `$ make`, or the considerably faster
parallel build `$ make -j<threads in machine + 1>`. This will create a
build folder and *copy* all files to their correct places within
this folder (once). This way symlinks and other loose items are
prevented within the llvm and clang git repos. It then triggers
cmake, followed by make, building `skepu-tool`.