Added skepu_add_executable function to cmake scripts.
The function adds targets to precompile SkePU sources when adding an executable
target. It also checks for MPI, OpenMP, and OpenCL so that the correct CXXFLAGS
is generated by cmake. It cannot check for CUDA, so we leave that to the user.
The user guide should reflect this. It cannot check for CUDA, so we leave that
to the user. The user guide should reflect this.